fix(issue-views): Display typed filters correctly in new issue view page
