+++ static/app/views/issueList/utils/useIssueListFilterKeys.tsx
@@ -1,0 +1,23 @@
+import {getUtcDateString} from 'sentry/utils/dates';
+import useOrganization from 'sentry/utils/useOrganization';
+import usePageFilters from 'sentry/utils/usePageFilters';
+import {useFetchIssueTags} from 'sentry/views/issueList/utils/useFetchIssueTags';
+
+export function useIssueListFilterKeys() {
+  const organization = useOrganization();
+  const {selection: pageFilters} = usePageFilters();
+  const {tags: issueTags} = useFetchIssueTags({
+    org: organization,
+    projectIds: pageFilters.projects.map(id => id.toString()),
+    keepPreviousData: true,
+    start: pageFilters.datetime.start
+      ? getUtcDateString(pageFilters.datetime.start)
+      : undefined,
+    end: pageFilters.datetime.end
+      ? getUtcDateString(pageFilters.datetime.end)
+      : undefined,
+    statsPeriod: pageFilters.datetime.period,
+  });
+
+  return issueTags;
+}
